requirements.txt
langchain-google-genai 
langgraph 
pandas 
streamlit 
plotly 
seaborn 
matplotlib 
fpdf2 
openpyxl
python-dotenv
langchain-ollama